backend/requirements.txt
requests
python-multipart
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.3.0
click==8.2.1
fastapi
uvicorn[standard]
greenlet==3.2.4
h11==0.16.0
idna==3.10
load-dotenv==0.1.0
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
pydantic[email]
boto3
sqlalchemy
psycopg2-binary
python-jose
email-validator
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
